Flagify turning off non-Set-Cookie events to save on bandwidth for the heatmap integration
