+++ lib/cookiemonster.js
@@ -23,6 +23,8 @@ let cookiePermSvc = Cc["@mozilla.org/cookie/permission;1"]
 let simplePrefs = require("simple-prefs").prefs;
 
 const kSTUDY_NAME = "cookiemonster";
+exports.kSTUDY_NAME = kSTUDY_NAME;
+
 //const kPULSE_INTERVAL = 24 * 60 * 60 * 1000; // One day in milliseconds
 const kPULSE_INTERVAL = 60 * 1000; // One day in milliseconds
 const kSTUDY_DURATION = 7 * kPULSE_INTERVAL; // One week
@@ -374,11 +376,16 @@ function recordEvent(event) {
   events.emit(kSTUDY_NAME, {subject: null, data: eventStr});
 }
 
-// Register all observer event handlers
-events.on("http-on-modify-request", onModifyRequest, false);
+// Register all observer event handlers. onExamineResponse is for Set-Cookie
+// events, the most important type.
 events.on("http-on-examine-response", onExamineResponse, false);
-events.on("quit-application", onQuitApplication, false);
-events.on("cookie-changed", onCookieChanged, false);
+
+// Register the rest of the handlers, if we aren't bandwidth-constrained.
+if (simplePrefs.enable_all_events) {
+  events.on("http-on-modify-request", onModifyRequest, false);
+  events.on("quit-application", onQuitApplication, false);
+  events.on("cookie-changed", onCookieChanged, false);
+}
 
 // We only export this for testing
 // Return a promise that resolves when all of the prefs are recorded.
@@ -406,40 +413,6 @@ function recordMetadata() {
                         });
     });
 }
-
-/**
- * Schedule data upload every 24 hours
- * @returns void
- **/
-function scheduleUpload() {
-  if (!storage.lastUpload) {
-    storage.lastUpload = Date.now();
-  }
-
-  let simulate = simplePrefs.enable_reporting;
-  // Write cookiemonster collected data to the MP study.
-  fuse({ start: storage.lastUpload,
-         duration: kSTUDY_DURATION,
-         pulseinterval: kPULSE_INTERVAL,
-         pulsefn: function _upload() {
-           // Record any metadata to accompany this upload event
-           recordMetadata().
-           // Upload the collected data
-           then(function() {
-             storage.lastUpload = Date.now();
-             return monitor.upload(kUPLOAD_URL, {simulate: fakeUpload});
-           }).
-           // Clear the monitor
-           then(function _clear(response) {
-             console.log("upload response", JSON.stringify(response));
-             // return monitor.clear();
-           });
-         },
-         resolve_this: self
-       });
-}
-
-scheduleUpload();
 
 /**
  * Check channel to see if it is part of a private browsing window
